+++ docs/cookbook/recipe02.py
@@ -1,4 +1,3 @@
-from __future__ import print_function  # necessary only if using Python 2.7
 from taxcalc import *
 
 # use publicly-available CPS input file