+++ taxcalc/calculate.py
@@ -7,13 +7,10 @@ Tax-Calculator federal tax Calculator class.
 #
 # pylint: disable=invalid-name,no-value-for-parameter,too-many-lines
 
-from __future__ import print_function
 import os
-import sys
 import json
 import re
 import copy
-import six
 import numpy as np
 import pandas as pd
 from taxcalc.functions import (TaxInc, SchXYZTax, GainsTax, AGIsurtax,
@@ -149,13 +146,6 @@ class Calculator(object):
                 print('Tax-Calculator startup automatically ' +
                       'extrapolated your data to ' +
                       str(self.__records.current_year) + '.')
-        if verbose and sys.version_info.major == 2:  # running Python 2.7
-            print(  # pragma: no cover
-                ('WARNING: Tax-Calculator packages for Python 2.7 will\n'
-                 '         no longer be provided beginning in 2019\n'
-                 '         because Pandas is stopping development for 2.7\n'
-                 'SOLUTION: upgrade to Python 3.6 now')
-            )
         assert self.__policy.current_year == self.__records.current_year
         self.__stored_records = None
 
@@ -1134,7 +1124,7 @@ class Calculator(object):
     def read_json_param_objects(reform, assump):
         """
         Read JSON reform and assump objects and
-        return a single dictionary containing six key:dict pairs:
+        return a single dictionary containing 6 key:dict pairs:
         'policy':dict, 'consumption':dict, 'behavior':dict,
         'growdiff_baseline':dict, 'growdiff_response':dict, and
         'growmodel':dict.
@@ -1169,7 +1159,7 @@ class Calculator(object):
             gdiff_base_dict = dict()
             gdiff_resp_dict = dict()
             growmodel_dict = dict()
-        elif isinstance(assump, six.string_types):
+        elif isinstance(assump, str):
             if os.path.isfile(assump):
                 txt = open(assump, 'r').read()
             else:
@@ -1184,7 +1174,7 @@ class Calculator(object):
         # next process first reform parameter
         if reform is None:
             rpol_dict = dict()
-        elif isinstance(reform, six.string_types):
+        elif isinstance(reform, str):
             if os.path.isfile(reform):
                 txt = open(reform, 'r').read()
             else:
@@ -1519,6 +1509,7 @@ class Calculator(object):
         string parameters as secondary keys.  This returned dictionary is
         suitable as the argument to the Policy implement_reform(prdict) method.
         """
+        # pylint: disable=too-many-locals
         # strip out //-comments without changing line numbers
         json_str = re.sub('//.*', ' ', text_string)
         # convert JSON text into a Python dictionary
@@ -1653,7 +1644,7 @@ class Calculator(object):
         # optionally convert lists into np.arrays
         year_param = dict()
         for pkey, sdict in param_key_dict.items():
-            if not isinstance(pkey, six.string_types):
+            if not isinstance(pkey, str):
                 msg = 'pkey {} in reform is not a string'
                 raise ValueError(msg.format(pkey))
             rdict = dict()
@@ -1661,7 +1652,7 @@ class Calculator(object):
                 msg = 'pkey {} in reform is not paired with a dict'
                 raise ValueError(msg.format(pkey))
             for skey, val in sdict.items():
-                if not isinstance(skey, six.string_types):
+                if not isinstance(skey, str):
                     msg = 'skey {} in reform is not a string'
                     raise ValueError(msg.format(skey))
                 else: